src/csv_db_package/requirements.txt
pandas~=1.3.5
mysql-connector-python~=8.0.27
Jinja2~=3.0.3